refactor global.d.ts
