+++ src/@types/assets.d.ts
@@ -1,0 +1,24 @@
+declare module '*.png' {
+  const value: string;
+  export default value;
+}
+
+declare module '*.jpg' {
+  const value: string;
+  export default value;
+}
+
+declare module '*.svg' {
+  const value: string;
+  export default value;
+}
+
+declare module '*.scss' {
+  const content: any;
+  export default content;
+}
+
+declare module '*.css' {
+  const content: any;
+  export default content;
+}
